Added TFL Definitions from CSV and  DB insertions method

+++ src/main/scala/com/PredictionAlgorithm/Database/DatabaseDocuments.scala
@@ -5,7 +5,7 @@ sealed trait DatabaseDocuments {
 
 }
 
-final case class POINT_TO_POINT_DOCUMENT(val route_ID: String, val direction_ID: Int, val from_Point_ID: String, val to_Point_ID: String, val day_Type: String, val dep_Time: Int, val duration: Long, val insert_TimeStamp: Long) extends DatabaseDocuments {
+final case class POINT_TO_POINT_DOCUMENT(val route_ID: String, val direction_ID: Int, val from_Point_ID: String, val to_Point_ID: String, val day_Type: String, val dep_Time: Long, val duration: Long, val insert_TimeStamp: Long) extends DatabaseDocuments {
 
   override val collection = POINT_TO_POINT_COLLECTION
